[ REFACTOR ] Apollo-server app.js 의 socket로직 리팩토링 (#81)
Why
- app.js가 방대해져서 분리를 해야 한다.

What
- socket 로직을 분리해서 webSocket폴더로 이동.
- 유저를 카운트 하는 부분을 빼내서 관리.
- 채널 스피커의 입장을 확인하는 로직을 뺴내서 관리.
